+++ shallow_backup/config.py
@@ -1,16 +1,23 @@
 import sys
 import json
+import os
 from os import path, environ
 from .printing import *
 from .compatibility import *
-from .utils import safe_mkdir
+from .utils import safe_mkdir, strip_home
+from .constants import ProjInfo
 
 
-def get_xdg_config_path():
+def get_xdg_config_path() -> str:
+	"""Returns path to $XDG_CONFIG_HOME, or ~/.config, if it doesn't exist."""
 	return environ.get('XDG_CONFIG_HOME') or path.join(path.expanduser('~'), '.config')
 
 
-def get_config_path():
+def get_config_path() -> str:
+	"""
+	Detects if in testing or prod env, and returns the right config path.
+	:return: Path to config.
+	"""
 	test_config_path = environ.get('SHALLOW_BACKUP_TEST_CONFIG_PATH', None)
 	if test_config_path:
 		return test_config_path
@@ -18,62 +25,94 @@ def get_config_path():
 		return path.join(get_xdg_config_path(), "shallow-backup.conf")
 
 
-def get_config():
+def get_config() -> dict:
 	"""
-	Returns the config.
-	:return: dictionary for config
+	:return Config.
 	"""
 	config_path = get_config_path()
-	with open(config_path) as f:
+	with open(config_path) as file:
 		try:
-			config = json.load(f)
+			config = json.load(file)
 		except json.decoder.JSONDecodeError:
 			print_red_bold(f"ERROR: Invalid syntax in {config_path}")
 			sys.exit(1)
 	return config
 
 
-def write_config(config):
+def write_config(config) -> None:
 	"""
 	Write to config file
 	"""
-	with open(get_config_path(), 'w') as f:
-		json.dump(config, f, indent=4)
+	with open(get_config_path(), 'w') as file:
+		json.dump(config, file, indent=4)
 
 
-def get_default_config():
-	"""
-	Returns a default, platform specific config.
-	"""
+def get_default_config() -> dict:
+	"""Returns a default, platform specific config."""
 	return {
-		"backup_path"      : "~/shallow-backup",
-		"dotfiles"         : [
-			".bashrc",
-			".bash_profile",
-			".gitconfig",
-			".profile",
-			".pypirc",
-			f"{get_config_path}",
-			".tmux.conf",
-			".vimrc",
-			".zlogin",
-			".zprofile",
-			".zshrc"
-		],
-		"dotfolders"       : [
-			".ssh",
-			".vim"
-		],
-		"default-gitignore": [
+		"backup_path": "~/shallow-backup",
+		"dotfiles": {
+			".bash_profile": {
+				"backup_condition": "",
+				"reinstall_condition": "",
+			},
+			".bashrc": {
+				"backup_condition": "",
+				"reinstall_condition": "",
+			},
+			".config/git": {
+				"backup_condition": "",
+				"reinstall_condition": "",
+			},
+			".config/nvim/init.vim": {
+				"backup_condition": "",
+				"reinstall_condition": "",
+			},
+			".config/tmux": {
+				"backup_condition": "",
+				"reinstall_condition": "",
+			},
+			".config/zsh": {
+				"backup_condition": "",
+				"reinstall_condition": "",
+			},
+			".profile": {
+				"backup_condition": "",
+				"reinstall_condition": "",
+			},
+			".pypirc": {
+				"backup_condition": "",
+				"reinstall_condition": "",
+			},
+			".ssh": {
+				"backup_condition": "",
+				"reinstall_condition": "",
+			},
+			".zshenv": {
+				"backup_condition": "",
+				"reinstall_condition": "",
+			},
+			f"{strip_home(get_config_path())}": {
+				"backup_condition": "",
+				"reinstall_condition": "",
+			},
+		},
+		"root-gitignore": [
 			"dotfiles/.ssh",
 			"dotfiles/.pypirc",
 			".DS_Store"
 		],
-		"config_mapping"   : get_config_paths()
+		"dotfiles-gitignore": [
+			".ssh",
+			".pypirc",
+			".DS_Store",
+		],
+		"config_mapping": get_config_paths(),
+		"lowest_supported_version": ProjInfo.VERSION
 	}
 
 
-def safe_create_config():
+def safe_create_config() -> None:
 	"""
 	Creates config file if it doesn't exist already. Prompts to update
 	it if an outdated version is detected.
@@ -85,24 +124,10 @@ def safe_create_config():
 		backup_config = get_default_config()
 		safe_mkdir(os.path.split(backup_config_path)[0])
 		write_config(backup_config)
-	else:
-		# If it does exist, make sure it's not outdated.
-		# TODO: Move this to upgrade.py
-		with open(backup_config_path) as config:
-			if "[USER]" in config.readline().strip():
-				if prompt_yes_no("An outdated config file has been detected. Would you like to update this?",
-				                 Fore.GREEN):
-					delete_config_file()
-					safe_create_config()
-				else:
-					print_red_bold("ERROR: Outdated config file found.")
-					sys.exit(0)
 
 
-def delete_config_file():
-	"""
-	Deletes config file.
-	"""
+def delete_config_file() -> None:
+	"""Delete config file."""
 	config_path = get_config_path()
 	if os.path.isfile(config_path):
 		print_red_bold("Deleting config file.")
@@ -111,24 +136,57 @@ def delete_config_file():
 		print_red_bold("ERROR: No config file found.")
 
 
+def add_dot_path_to_config(backup_config: dict, file_path: str) -> dict:
+	"""
+	Add dotfile to config with default reinstall and backup conditions.
+	Exit if the file_path parameter is invalid.
+	:backup_config: dict representing current config
+	:file_path:		str  relative or absolute path of file to add to config
+	:return new backup config
+	"""
+	abs_path = path.abspath(file_path)
+	if not path.exists(abs_path):
+		print_path_red("Invalid file path:", abs_path)
+		return backup_config
+	else:
+		stripped_home_path = strip_home(abs_path)
+		print_path_blue("Added:", stripped_home_path)
+		backup_config["dotfiles"][stripped_home_path] = {
+			"reinstall_condition": "",
+			"backup_condition": ""
+		}
+	return backup_config
+
+
 def show_config():
 	"""
 	Print the config. Colorize section titles and indent contents.
 	"""
 	print_section_header("SHALLOW BACKUP CONFIG", Fore.RED)
 	for section, contents in get_config().items():
-		# Hide gitignore config
-		if section == "default-gitignore":
-			continue
 		# Print backup path on same line
-		elif section == "backup_path":
+		if section == "backup_path":
 			print_path_red("Backup Path:", contents)
 		elif section == "config_mapping":
-			print_red_bold("Configs:")
+			print_red_bold("\nConfigs:")
 			for path, dest in contents.items():
-				print("    {} -> {}".format(path, dest))
-		# Print section header and intent contents. (Dotfiles/folders)
+				print(f"	{path} -> {dest}")
+		# Print section header and contents. (Dotfiles)
+		elif section == "dotfiles":
+			print_path_red("\nDotfiles:", "(Backup and Reinstall conditions will be shown if they exist)")
+			for dotfile, options in contents.items():
+
+				backup_condition = options['backup_condition']
+				reinstall_condition = options['reinstall_condition']
+				if backup_condition or reinstall_condition:
+					print(f"	{dotfile} ->")
+					print(f"\t\tbackup_condition: \"{backup_condition}\"")
+					print(f"\t\treinstall_condition: \"{reinstall_condition}\"")
+				else:
+					print(f"	{dotfile}")
+		elif section == "lowest_supported_version":
+			print_path_red(f"{section.replace('_', ' ').capitalize()}:", contents)
 		else:
-			print_red_bold("\n{}: ".format(section.capitalize()))
+			print_red_bold(f"\n{section.replace('-', ' ').capitalize()}: ")
 			for item in contents:
-				print("    {}".format(item))
+				print(f"	{item}")