+++ shallow_backup/printing.py
@@ -1,3 +1,4 @@
+import os
 import sys
 import inquirer
 from colorama import Fore, Style
@@ -52,6 +53,40 @@ def print_path_green(text, path):
 	print(Fore.GREEN + Style.BRIGHT + text, Style.NORMAL + path + Style.RESET_ALL)
 
 
+def print_dry_run_copy_info(source, dest):
+	"""Show source -> dest copy. Replaces expanded ~ with ~ if it's at the beginning of paths.
+	source and dest are trimmed in the middle if needed. Removed characters will be replaced by ...
+	:param source: Can be of type str or Path
+	:param dest: Can be of type str or Path
+	"""
+	def shorten_home(path):
+		expanded_home = os.path.expanduser("~")
+		path = str(path)
+		if path.startswith(expanded_home):
+			return path.replace(expanded_home, "~")
+		return path
+
+	def truncate_middle(path: str, acceptable_len: int):
+		"""Middle truncate a string
+		https://www.xormedia.com/string-truncate-middle-with-ellipsis/
+		"""
+		if len(path) <= acceptable_len:
+			return path
+		# half of the size, minus the 3 .'s
+		n_2 = int(acceptable_len / 2 - 3)
+		# whatever's left
+		n_1 = int(acceptable_len - n_2 - 3)
+		return f"{path[:n_1]}...{path[-n_2:]}"
+
+	trimmed_source = shorten_home(source)
+	trimmed_dest = shorten_home(dest)
+	longest_allowed_path_len = 87
+	if len(trimmed_source) + len(trimmed_dest) > longest_allowed_path_len:
+		trimmed_source = truncate_middle(trimmed_source, longest_allowed_path_len)
+		trimmed_dest = truncate_middle(trimmed_dest, longest_allowed_path_len)
+	print(Fore.YELLOW + Style.BRIGHT + trimmed_source + Style.NORMAL, "->", Style.BRIGHT + trimmed_dest + Style.RESET_ALL)
+
+
 def print_version_info(cli=True):
 	"""
 	Formats version differently for CLI and splash screen.
@@ -66,17 +101,13 @@ def print_version_info(cli=True):
 
 
 def splash_screen():
-	"""
-	Display splash graphic, and then stylized version and author info.
-	"""
+	"""Display splash graphic, and then stylized version and author info."""
 	print(Fore.YELLOW + Style.BRIGHT + "\n" + ProjInfo.LOGO + Style.RESET_ALL)
 	print_version_info(False)
 
 
 def print_section_header(title, color):
-	"""
-	Prints variable sized section header
-	"""
+	"""Prints variable sized section header."""
 	block = "#" * (len(title) + 2)
 	print("\n" + color + Style.BRIGHT + block)
 	print("#", title)
@@ -93,6 +124,7 @@ def print_pkg_mgr_reinstall(mgr):
 												  mgr, Fore.BLUE, Style.NORMAL, Style.RESET_ALL))
 
 
+# TODO: BUG: Why does moving this to prompts.py cause circular imports?
 def prompt_yes_no(message, color, invert=False):
 	"""
 	Print question and return True or False depending on user selection from list.
@@ -108,3 +140,5 @@ def prompt_yes_no(message, color, invert=False):
 		return answers.get('choice').strip().lower() == 'yes'
 	else:
 		sys.exit(1)
+
+