+++ shallow_backup/prompts.py
@@ -1,12 +1,11 @@
 import os
 import sys
 import inquirer
-from colorama import Fore, Style
 from .utils import *
 from .printing import *
 from .config import *
 from .git_wrapper import git_set_remote, move_git_repo
-from .utils import new_dir_is_valid
+from .utils import check_if_path_is_valid_dir
 
 
 def path_update_prompt(config):
@@ -16,12 +15,12 @@ def path_update_prompt(config):
 	"""
 	current_path = config["backup_path"]
 	print_path_blue("Current shallow-backup path:", current_path)
-	if prompt_yes_no("Would you like to update this?", Fore.GREEN, invert=True):
+	if prompt_yes_no("Would you like to move this somewhere else?", Fore.GREEN, invert=True):
 		while True:
 			print_green_bold("Enter relative or absolute path:")
 			abs_path = expand_to_abs_path(input())
 
-			if not new_dir_is_valid(abs_path):
+			if not check_if_path_is_valid_dir(abs_path):
 				continue
 
 			print_path_blue("\nUpdating shallow-backup path to:", abs_path)